// src/types.ts
export interface Employee {
  stt: string;
  msnv: string;
  avatarUrl: string;
  fullName: string;
  birthDate: string;
  phone: string;
  email: string;
  gender: string;
  deptCode: string;
  deptShort: string;
  department: string;
  team: string;
  roleCode: string;
  role: string;
  educationLevel: string;
  educationMajor: string;
  partyBranch: string;
  partyMember: string;
  unionMember: string;
  youthUnionMember: string;
  base: string;
  oldBase: string;
  address: string;
}